Revert the reversion of r304 since it doesn't seem to crash anymore
Original commit was
Reset apps scope when returning from app to dash (LP #1193419)

If an app is on foreground and you perform a long left-edge swipe to minimize it,
and therefore return to the dash, the dash should be in the Applications scope
and showing the running/recents applications. Fixes: https://bugs.launchpad.net/bugs/1193419.