Refactor Ubuntu.Gesture tests to share common logic
Take the common part out of tst_DirectionalDragArea and put
it into a separate base class, GestureTest, so that it can
be shared with other, future, tests.

In CMakeLists.txt, create a macro out of DirectionalDragArea build script
to be used by future tests with similar requirements and structure.

Also add the "m_" prefix to member variables.