unlock_device: support more complex reboot/wait cycles
Currently this script only allows you to override how to "wait" on the
device. This changes the logic to also support how you go about rebooting
the device.

This is handy for the ubuntu-emulator because adb-reboot is not
currently supported. However, we also have a more sophisticated,
fool-proof way we reboot/wait in the CI lab that would be nice
to take advantage of:

 http://bazaar.launchpad.net/~ubuntu-test-case-dev/ubuntu-test-cases/touch/view/head:/scripts/reboot-and-wait