Desperate hack to fix QML error
Qt was complaining that it was unable to assign to a inexistent property
on Shell.qml:299 and thus was unable to load the Shell component.
That's the "background: shell.background" line in the Lockscreen {...} code
in Shell.qml. But that error just doesn't make sense as Lockscreen does define
a background property. Getting this error on an N4.

I've no idea of what's causing this error but at least commenting out this line
makes it go away with seemingly no side-effects.