Add touch ownership logic on top of qt input handling
It does mainly 3 things:
- Adds TouchRegistry.
- Adds TouchGate
- Modifies DirectionalDragArea so that it uses TouchRegistry and only grabs a touch once it has recognized a gesture done by it.

TouchRegistry was put into a library as both unity8 and Ubuntu.Gestures plugin need to use it.

The next step is to have surfaces participating in the touch ownership negotiation. But for now they're put behind TouchGates so that they only receive thouches whose ownership has already been settled.

The most noticeable improvement of this patch is that taps near screen edges will now reach applications.