Make TouchGate synthesize QMouseEvents for mouse-based target items
As the dispatch code grew considerably because of that it was moved out of
TouchGate, into a separate object called TouchDispatcher.

Moved some common test code out of tst_DirectionalDragArea.

Also added TouchGate to cover Greeter's MouseArea, so that it only triggers
hint animations and tease launcher once we are sure that a left-edge
drag won't take place. ie: don't tease launcher while it's being dragged
at the same time.