improve snapping after dragging
This fixes an issue that the listView was doing some weird bounce effect when dropping
items at the top or bottom because of some flicking that tried to make sure the dropped
item is completely visible. This didn't work reliably and introduced the above issue.
The new code uses some animation to move around contentY which seems to be much more
reliable and avoids the issue.