tests: Fix Notify.Notification.add_action invocation
Drop unnecessary "free_func" argument from the Notification add_action() call.
This only works because of an Ubuntu specific libnotify patch which exposes a
broken and old introspection ABI, but we want to get rid of this.

This will now work with both the upstream *and* the old Ubuntu API, therefore
this does not regress anywhere.