Set currentIndex after the model change is completed
When setting the model, the list view will first emit the countChanged signal,
then will cahnge the currentIndex to 0 (see QtDeclarative git commit 22c6873
and https://bugreports.qt.io/browse/QTBUG-26604), and finally will emit the
modelChanged signal.
If we set the currentIndex in the onCountChanged handler (as we were doing
before), the currentIndex would be reset to 0 shortly afterwards.