Keep the PreviewStack around
The PreviewStack ownership is handled from the C++ plugin to QML so we need to keep it around
otherwise it may be deleted (depends on the garbage collection) when we're still using the preview models
that are its child and they'll be garbage collected too.

This did not happen yet because the plugin had a bug that caused the preview stacks to never be deleted.