Fix test failing on Xenial (Qt 5.5)
This is the simplest of the ways of fixing the problem i found.

The issue is that flickable code doesn't play very well with the situation in which you have a contentHeight smaller than height
and you reimplement minYExtent but not maxYExtent and minYExtent is returning a negative value.

The other solutions involve implementing maxYExtent figuring out why minYExtent ends up being negative on Qt 5.5 and not in Qt 5.4

The first seemed too much work

The second is most probably ok since it's just different order invokations lead to different minyExtent+friends combinations that basically
represent the same thing

So I ended up making sure the minYExtent is never negative (because it's 0) if the contentHeight is smaller than the height of the list