improve the pin lockscreen
On smallish screens, like the BQ E4.5, it frequently happens that one
misses a digit when quickly entering the pin. After a while of debugging
usage on my phone I noticed that the cause of this is actually that the
user keeps moving the finger while touching the buttons and so it
often happens that the touch events are not recognized as clicks.

I've been using this patch for a while now and it increased the reliability
of the pin entry a lot. I've also asked other people who reported the issue
to test this and feedback has been rather good.