Workaround QTBUG-53460
Make the invisible dummy highlight animate fast as to not cause extra rendering                                                                                                                                                                                           
      
The preview change is easy to reproduce, the gallery one is a bit harder but is based on the same principle.
      
For the preview one, open the Telegram app on the store and scroll down quite a bit and leave it there.
Observe how both unity8-dash and unity8 use around 20% CPU (on the phone) for a LONG time
      
Now do the same with this patch and see the CPU usage drops to 0 quickly.

The same workaround could apply to other ListViews but i can not reproduce the behaviour there so i was cautious