Deprecate usage of Mir's input resampling, instead opting for:   Shell: Raw native input events   Apps:  Qt's built-in touch compression input resampling algorithm,          and other toolkits are expected to provide their own.
This fixes LP: #1497105 nicely, allowing apps to receive the full-speed
input stream (if they ask their toolkit to give it to them).

As an added bonus this also eliminates the 0-16.9ms (average 8.4ms) of
input lag that apps under Unity8 would experience. We do still have the
problem of output lag making everything still look laggy but I am
working to eliminate that soon in Mir 0.26.0.