[Stage/Spread] Detect mouse or touch events in newer qt versions
With newer versions of qt a touch event will no longer set "pressed" to
true, this is what the spread has been relying on to detect mouse or
touch events. This adds detection if it's a mouse or touch event in
never versions of qt.

This fixes: https://github.com/ubports/ubuntu-touch/issues/487 and
https://github.com/ubports/ubuntu-touch/issues/430