Add a "Welcome back" wizard for Xenial
That's an understatement. This commit adds the ability to create
"versioned" wizards, where the Wizard will re-run with certain pages
the next time that Unity starts. Creating a new Wizard version is as
easy as updating the CURRENT_VERSION in plugins/Wizard/System.h. Each
Page in the Pages folder will run only on the first-run wizard by
default, but by setting showOnVersions to an empty array or a list of
specific versions, they will run on the "update" Wizards as well.
Welcome, Wi-Fi, and Finished have been set to run on all "update" runs
of the wizard. I've also added an "Update apps" page which forces a
check of all apps. Perfect for Xenial where all apps should be
reinstalled cleanly from the store.

I've also added a spinner which shows when pages are taking a while to
load. This isn't as good as fixing the loading issues, but I think a
complete refactor of the Wizard would be in order to achieve that.