Update README.md
I notice that many are reporting issues with Unity8 in the wrong issue tracker, namely on the Unity8 installer issue tracker, probably because that's where the installation instructions are found.   Moving the installation instructions to this readme (and pointing the readme found in the installer repository to here) may reduce this problem moving forward.