Fix interaction papercuts with the Wizard
There are a couple of small issues when interacting with the 16.04
update wizard. The first is that the app update page Stop button is
ineffective. My solution to this is removing it altogether since
stopping the updater in the wizard could cause user confusion later when
their apps don't work correctly. Also, after hours of troubleshooting
I'm still not sure why the check won't move to another thread like in
its home, system-settings.

Another problem is that the flow to unlock a SIM is not available in the
Wizard, so the app update page may not be able to connect to the
internet. Instead of implementing the SIM unlock in the Wizard, I've
made it clear that updates can be found in the OpenStore app after the
device has been connected.
See: https://github.com/ubports/ubuntu-touch/issues/842