Remove the swipe to close the Drawer :(
In order to avoid glitchy MouseArea inside Flickable behavior (events
not getting delivered or geing created when they shouldn't), remove the
MouseArea!
Unfortunately, this means that we lose the ability to drag the Drawer
away... We can bring it back later when we know how to implement it
better, but for now we'll use the BFB as a Back button when the Drawer
is pulled out.