Squash race conditions everywhere
I searched the entire codebase for compare() entries and evaluated all
of them for how appropriate they are compared to tryCompare() or another
alternative. I found that there were many cases where compare()
depended on delays that came before it to get the correct value, causing
race conditions that we haven't even hit yet. In a proactive measure,
this commit:

* Replaces compare() with tryCompare() where appropriate
* Replaces tryCompareFunction() with tryVerify() where appropriate
* Uses SignalSpy.wait() where appropriate
* Generally lints and pays back technical debt when I found the
  opportunity