Make sure we don't halt shutdown/reboot wating for apps to close
This was terrible to begin with, it relied on that the app would close
without problems and got popped from the array.

There is no api for getting all apps, so a hacky workaround for this is
just to check next index number for an app, and repeat this 20 times, if
the apps still wont close, just continue shutdown.

This fixes: https://github.com/ubports/ubuntu-touch/issues/1111