Fix display of the default wallpaper in the Drawer
It seems that our default setting for AccountsService.backgroundFile
references a file that doesn't exist any more. That's fine since the
default wallpaper is now handled differently in Unity8. However, we
needed to extend that handling to the Drawer.

This commit switches from using AccountsService.backgroundFile as our
wallpaper source to the cached scaled-down wallpaper held by
wallpaperResolver in Shell.qml. Since we're blurring the wallpaper
anyway, we can use this scaled-down version and save some memory.