Make sure no apps have focus if stage is not active/focused
Make sure no apps have focus if stage is not active/focused, to achive
this a null window is used to steal focus from the other apps.

The reason why a null window is used is that we now use miral to handle
all focusing of apps, and since thats only aware of clients it wont know
if unity8 itself wants the focus. So to workaound this we make a null
window have the focus while unitu8 have focus.

Fixes: https://github.com/ubports/unity8/issues/130
Fixes: https://github.com/ubports/ubuntu-touch/issues/1123