Refactor handling of mainApp changes
While testing for an unrelated change, I noticed that it was possible to
put Unity8 in a bad state if the dialer crashed during an emergency
call. This wasn't a security issue since the Shell still knew it was
supposed to lock everything down and blocked access to indicators and
launching apps (for example).

This became a good chance to learn how the locked app support works and
how Unity8 generally decides how to focus or unfocus certain things in
response to apps starting. I took the chance to refactor it a bit.

* Move the "show Launcher when no apps are running" logic to run when
  the focused app changes. Cuts down a Connection.
* *Always* call _onMainAppChanged to ensure the Greeter knows about the
  focused app
* Refactor and update comments within _onMainAppChanged