Let TopLevelWindowModel deside what app to refocus, if any
This let TopLevelWindowModel deside what app it should refocus after a
the null window has been focused. It will only refocus the previous
window if null focus still has focus. This makes sure we dont give the
focus to another app while an app is starting, or other window has
clamed the focus.

Pending Activation will block refocus of previous focused window
this is needed since a surface activation with miral is async
and a activation of placeholder is sync. This causes a race condition
between placeholder and prev window. This results in prev window
gets focused as it will always be later then placeholder.

This fixes: https://github.com/ubports/ubuntu-touch/issues/1170