Work around greeter anchoring mess (#190)
This hides the user manager to work around #185. Instead, it will display a new label that's properly anchored.

Yes. I know: It's not pretty. But there's a lot of unused complexity here that we might need later. I don't want to touch all that now, but i am also convinced that we need to fix this. Every person i give my M10 to notices this, and that is not the first impression i want people to get of Ubuntu Touch. At some point we will have to refactor this and fix it properly.