Enable titleLabel in portrait mode too.
This is meant as a design suggestion.

I like the window title that shows up when using my phone in landscape and wanted it to display all the time. I've been using the Panel this way for a couple months and have noticed no ill effects. It won't let me upload an image to show what it llike.like so I will try after posting this.

Perhaps it could also be an option in the settings menu of the phone, but I do not know how to implement such.