Make the appDelegate know when it's being focused
There's a "slide" animation that plays when apps are focused on their
own request or on the request of an outside party.

Unfortunately, this animation plays after the rightEdgeFocusAnimation,
which calls 'appDelegate.activate()' when it stops. When that happens,
the delegate acts like it was focused by another source and plays the
sliding animation. This behavior was probably introduced in f1dcc66.

The inhibitSlideAnimation property of each appDelegate gets set when it
is focused with the flip animation and unset when it becomes unset. This
strikes a good balance between simplicity and ensuring all the
functionality (sliding in and out when another app is focused, for
example) still works.

Also, ensure the slide animations take the speed of the animations
imposed by the Stage.