Hide keyboard when Launcher or Drawer are dragged
The Launcher and Drawer do not respect the on-screen keyboard rectangle,
so about half of them are hidden when the Drawer's search field is
focused.

To fix this, unfocus the Drawer's search field when either of them move.

This does not remove focus when the user scrolls with the mouse wheel,
only when they click and drag or touch and drag.

Fixes https://github.com/ubports/ubuntu-touch/issues/1238
Fixes https://github.com/ubports/ubuntu-touch/issues/1245