[TLWM] Change rootFocus to a property
This changes rootFocus to a qt property insted of a callable function.
This makes the API make more sense in QML.

In addition we add some C++ tests for the rootFocus behavior. There are
already some roundabout tests for this behavior in the qmltests for
the different Stages, but this is more direct.