TIme out findChild faster when we expect null
The default timeout for findChild is 5 seconds. This timeout is reached
whenever the child is not found in the requested parent. If we expect
the child to be null anyway,  we don't need to wait that long. 1 second
is plenty of time to search the small children area of 'dialogs' anyway.

This saves about 40 seconds on my unloaded Ryzen 7 1700.