Add global InputMethodManager to handle inputMethod across screens
This adds a global InputMethodManager to handle inputMethod across
screens. This (c++ side) was made with to be comparable with the
upcoming proper multiscreen support.

This is a bit *hacky* for now, and should be cleaned up once proper
multiscreen has landed.

This is still less hacky then the existing hardcoded
multiscreen/secondscreen things currently in unity8 so ¯\_(ツ)_/¯

This fixes: https://github.com/ubports/ubuntu-touch/issues/1096