Make touchFlick flick instead of snap
Prior to this change, NaN was stored in 'distance' and every touchFlick
would snap the touch point from the start to the end instantly.

After fixing this, touchFlick took much longer, so I also increased the
default speed.

The only test that this broke was test_rightEdgePushWithOpenIndicators,
so I replaced the touchFlick with showIndicators(). While I was in the
file I left it a little nicer than I came by placing all utility
functions above the test functions.