Make it less likely to issue duplicate Window IDs
Some tests will issue the null window and the first window at almost the
same time, causing their window IDs to be the same prior to this change.
That was fun to debug.

I'm not going to pretend that this is absolutely foolproof, it's
possible we still end up with duplicate IDs, but the risk should be
quite small.

I had trouble testing this with the rest of the PhoneStage tests where
it is now customary to put the TLWM tests, there is too much troublesome
behavior when changing maxWindowId with an in-use model. Instead, the
new tst_QMLTopLevelWindowModel gives a case similar to PhoneStage but
with less touch input and more raw window actions.