Improve showing launcher automatically
If the Panel is open and the Greeter is swiped away, the Panel was
automatically dismissed by the Launcher appearing. Avoid this
(in most cases) by making sure the Panel isn't visible at all before
calling the Launcher.

A warning was generated in the tests when we tried to check
!greeter.active since greeter was null.