Refactor and fix intermittent tests
Some tests in tst_Shell would fail due to latent conditions, mostly due
to trying to open the Panel while the Launcher was animated. The
Launcher's animations block Panel actions.

To fix this, we can load the desktop stage with at least one app. This
prevents the Launcher from being called and helps the situation immensely.

While I was at it, I refactored the loadDesktopShellWithApps function to
allow loading fewer than the maximum number of apps, and to avoid using
unity8-dash since it has special behavior in the tests that we don't
want to test for any more. Then I used the refactored function
everywhere that we manually loaded the desktop shell before.

And for the cherry on top, I fixed a couple 'could not read property of
null' errors from the test controls.