Basically skip animations if Showable not shown
There are instances where a Showable may be told to hide even if it is
not currently shown. This happens quite a lot in use, since just saying
'hide()' is far better than 'if (shown) hide()'. However, this causes
the Showable's hideAnimation to play, even if it's already hidden.

This can cause an interesting problem when items' usability depends on
whether their animations are playing, such as the PanelMenus disabling
their DragHandles[1].

To fix this, immediately complete the hideAnimation (and showAnimation)
when they are not needed. Doing this instead of not playing the
animations at all corrects situations where shown was not set but the
Showable is being shown somehow.

This caused the panel tests to fail on timing issues (attempting to
start a touch event while the Panel was still disabled), so I had to
refactor one and add an ugly wait().

This also caused ShellWithPin's test_bfbOnLockedDevice to fail since the
shell tried to open the Drawer then immediately call launcher.hide().
Previously the timing worked out somehow and hide() was called first. To
fix this, remove the onSessionStarted() handler and move it to the
Greeter's onActiveChanged handler instead.

[1] https://github.com/ubports/unity8/pull/205