Keep refactoring tst_Shell, reduce race conditions
Refactor loadDesktopShellWithApps into loadShell to have a single place
to fix problems with shell or app loading in this test case.

Use the new refactored function everywhere.

I had to shuffle the order of setting up the greeter to cope with the
change, and add some more waitForRendering magic to make sure we don't
stomp on the Shell while it loads.