Don't be clever about wallpapers
It appears that our wallpaper rendering was a victim of premature
optimization. The image cache and cross-fading image caused some
significant performance and memory issues.

Qt includes its own cache for images. It's implicitly enabled whenever
you use an Image component. This cache ensures that only one copy of an
image is in RAM if possible. We explicitly disabled this on
CrossFadeImage[1].

The ImageCache also caused each image to be loaded from disk every time
it was used -- for example, every time the user locked their device and
the Greeter was created (twice for the NarrowView greeter!). This was
the cause of several noticeable pauses when navigating the shell as
well, for example when opening the Drawer. I've removed the ImageCache
as it is no longer needed.

Setting SourceSize in Wallpaper.qml actually increased memory use and
load times in my testing. For that reason I've removed the usage.

In my testing, this change saves up to 150MB of RAM (!) when sitting on
the narrow Greeter, and up to 50MB when sitting at the Background. The
numbers depend on the size of the wallpaper in use, but the reduction is
quite significant.

However, this change removes the background fading whenever it is
changed. I believe the performance benefits outweigh this eye candy.

[1] https://github.com/ubports/ubuntu-ui-toolkit/blob/eb36fcaad14e28b9e73fe4f4d5c73fbe81922855/src/imports/Components/1.2/CrossFadeImage10.qml#L94