When Launcher is locked, OSK can have full width
b62a87cd caused the Keyboard to appear at partial width when no apps
were running. This was clear when trying to use the Drawer's search or
when in the Wizard.

This commit refactors the Launcher's locking into clearer bindings and
uses those to determine whether the keyboard should have a margin or not

Fixes https://github.com/ubports/ubuntu-touch/issues/1373