Only allow tutorials to run serially
The only condition which each Tutorial relied on was the previous one
completing. This caused issues when, for example, the short left
tutorial was auto-skipped. When this occurred, both the top and
long-left tutorials played at once, locking the shell.

This commit makes each tutorial depend on the skipped state of all the
ones that came before it, avoiding this problem.