Allow indicators to take up more space
The addition of application menus made it more difficult to hit the
target area for the messages indicator. Prior to the application menu
addition, the entire blank area of the panel could be used. After, only
the 2 gu space taken up by the icon could be used.

This was due to an implementation detail of the application menus...
when the Shell has 60 gu or less width, both the application menus and
the indicators are set to the full width of the panel. This meant that
their drag handles, the area where they could be "pulled down" from,
took the entire panel. Since the indicators were defined after the
application menus in the file, the indicators' handle went over the
application menus' and the menus were unusable.

To solve this, the original author forced the indicators' handle to only
use the space taken by the indicator items. This approach had two big
problems. First, if the user had enough indicators to reach all the way
across the screen, the application menus would be unusable. Second, it
shrank the usable handle area for the leftmost indicator.

This change adjusts the indicator bar to provide a 7 grid unit space
to its left whenever the application menus are available. In case of
space contention between the two (read: too many indicators are
showing), the leftmost indicators will hide. Whenever the application
menus are not available, the indicators (and, more importantly, their
handle) take up the full width of the screen again.

Additionally, I added a "down" icon to the panel. It shows up at any
screen size, but only when application menus are available. This should
help improve discoverability of the application menus, which were
visually "silent" in Staged mode prior to this change.

Finally, I've noticed that sometimes tests may not clean up properly
after setting indicators to hidden. To fix this, I manually set the
indicators hidden then visible again. Since I needed a test case for
a smaller screen and didn't want to copy so much code, I refactored
the test case into a new PanelUI file for the setup.

Fixes https://github.com/ubports/ubuntu-touch/issues/1233