Compensate for PanelItemRow's origin changing
The QtQuick Flickable documentation [1] states that, "If you flick back
to the origin, after the rebound animation, contentX will settle to the
same value as originX. These are usually (0,0), however ListView and
GridView may have an arbitrary origin..."

This was causing problems for us in the Indicators. Some events, such as
a manual 'show()' of the PanelMenu, could cause the ListView's origin to
shift. When this occured, the highlight would be incorrect and other
interactions would not work correctly. For example, trying to select the
rightmost indicator would cause the leftmost to be selected.

To fix this, use the PanelItemRow's contentX to compensate some values
of X for the difference.

Fixes https://github.com/ubports/ubuntu-touch/issues/1364

[1] https://doc.qt.io/qt-5/qml-qtquick-flickable.html#contentX-prop