Add pull-to-refresh to app drawer
We have XdgWatcher that watches the change for Click apps. However,
there isn't such a thing for Libertine apps. This commit implement
pull-to-refresh in the app drawer as a stop-gap solution.

In the long term, when we have such mechanism, we can remove it. Or, we
can just leave it here as a failback when watcher things work
incorrectly.

Changelog:
  v2:
    - Switch to PullToRefresh component from Ubuntu.Components instead
      of homegrown solution. PullToRefreshScopeStyle is copied from the
      legacy Unity8 code.
    - Make sure AppDrawerModel clears the list before re-read. Add a
      test for that.
  v3:
    - Remove unnecessary change left over from v2.
    - Make sure all items in the list are deleted before being cleared.
  v4:
    - Completely re-implement AppDrawerModel to refresh asynchronously,
      including when first constructed. Expose a new "refreshing"
      property.
    - Display a progressbar when refresing is in progress.