d/control: temporary disable lomiri-system-settings
This is a dependency of the first-start wizard. However, currently
lomiri-system-settings is a monolith package that comes with most
plugins, and which in turn depend on indicators. We don't have
indicators yet, so comment this for now.

In the future, it might be worth it to split the wizard package out of
the main package, enabling the DE itself to run without the wizard.