PhysicalKeysMapper: handle first power button long press correctly
When a power button pressing is the very first input to Lomiri/Unity8,
the PhysicalKeysMapper won't recognize the long pressing correctly.

Turns out, the timestamp Qt gives us can start at 0, so using 0 as an
invalid value breaks this case. So, use -1 instead.

The test is updated to also test this case.