Prevent Drawer focus from flapping during transitions
I coded myself into a bit of a corner with the Drawer. Since it was
undesirable for the Maliit Copy/Paste dialog and text selection baubles
to stay rooted in place while the Drawer's search box moved underneath
them, we strategically saved and removed focus from the search bar. This
approach is very sensitive to changes in focus that come from outside
itself. The Launcher's States were one such area causing these
unexpected changes.

It appears that a State extending another state applies all
PropertyChanges from every state it extends. This meant that with the
"visible" state setting "focus: false" on the Drawer, then the "drawer"
state setting "focus: true", both changes occurred in quick succession.

To fix this rapid off-on cycle during transitions, change the "drawer"
state so it applies the relevant changes from "visible" but does not
actually extend "visible". We also need to disable "restoreEntryValues"
on the PropertyChanges because a "focus: false" before entering a state
does not necessarily mean that "focus: false" should be set when leaving
the same state.