Fix Greeter Wallpaper being positioned differently
When unlocking Lomiri, you would see a small jiggle effect as the
Greeter faded away. This was due to the Greeter's wallpaper being a
different size than the area below it. The Image.PreserveAspectCrop
aspect  would show a slightly different view of the image and the effect
came into being.

Fix this by extending the Greeter all the way to the top of the Shell.
Then, fix up the topMargin property so the narrowView's clock still
avoids the top of the screen.

The unit test tried to ensure that the Background would move up if the
Greeter was moved down with its y value. However, the topMargin value
was used to position the Greeter. anchors.topMargin does not modify the
item's y value. The test was therefore useless, so I removed it.